LayoutTests/imported/w3c/web-platform-tests/css/selectors/hover-002-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Hover me - should become green
Hover me - should become green

Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT Hover #hovered element should make it go green Test timed out
TIMEOUT Hover #hoveredContents child should make it go green Test timed out

Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT Hover #hovered element should make it go green Test timed out
TIMEOUT Hover #hoveredContents child should make it go green Test timed out
